feat: clickable comparison text
